pyproject.toml
[tool.poetry]
name = "djangoem-diplom"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
celery = "^5.4.0"
redis = "^5.0.6"
django-filter = "^24.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.7"
cors = "^1.0.1"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
black = "^24.4.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^70.1.0"
requests = "^2.32.3"
stripe = "^9.12.0"
coverage = "^7.5.4"
django-celery-beat = "^2.6.0"
eventlet = "^0.36.1"
gevent = "^24.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
